% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rXMLTV.R
\name{ReadXmltvFile}
\alias{ReadXmltvFile}
\title{Reads an XMLTV file}
\usage{
ReadXmltvFile(xmltv.file, verbose = FALSE)
}
\arguments{
\item{xmltv.file}{is the full path to the XMLTV file that should be read.}

\item{verbose}{If TRUE the functions prints the full path of the the
processed file. Defaults to FALSE.}
}
\description{
\code{ReadXmltvFile} reads one XMLTV file and returns its content in data
frame.
}
\details{
The XMLTV file can be tested before reading it by using
\code{\link{TestXmltvFile}}. The returned data frame inherits class
"xmltv.data.raw".
}

